refactor(FR-3104): restructure deployment detail into self-contained section cards (#7974)
Resolves #7839 (FR-3104)

> ⚠️ **Scope note:** This PR delivers the **structural refactor only**. The original FR-3104 goal — deduping the repeated `deployment(id:)` requests on the detail page via Relay refetchable fragments — is **blocked by [BA-6569](https://lablup.atlassian.net/browse/BA-6569)** (the backend's `node(id:)` does not resolve `ModelDeployment`, so `@refetchable` / `useRefetchableFragment` fails at runtime). Data fetching therefore stays as-is (per-tab lazy `useLazyLoadQuery(deployment(id:))`); the query dedup is **deferred until BA-6569 is fixed**.

## Summary

Restructures the Deployment detail page so every top-level section is a self-contained card, and decomposes the configuration/revision area into focused components. No data-fetching architecture changes — each tab keeps its existing lazy `deployment(id:)` query.

## Changes

- **Split `DeploymentConfigurationSection`** into two sibling cards:
  - `DeploymentBasicInfoCard` — overview + Edit / Delete / Scheduling-history.
  - `DeploymentRevisionCard` — thin tabbed card (`BAICard` + tabs only); each tab is its own component:
    - `DeploymentCurrentRevisionTab` (current revision detail + "applying" banner + rollout poll + revision drawer)
    - `DeploymentRevisionHistoryTab` (unchanged — a real sub-tab, lazy)
    - `DeploymentAuditLogTab` (audit log, loaded on tab activation)
- **`...Tab` → `...Card` renames** for the always-visible sections (+ matching GraphQL fragment / query / mutation names): `DeploymentReplicasCard`, `DeploymentAutoScalingCard`, `DeploymentAccessTokensCard`. `DeploymentRevisionHistoryTab` keeps its name (it is a genuine sub-tab).
- **AutoScaling split**: `AutoScalingRuleList` → presentational `AutoScalingRuleListNodes` (table) + `DeploymentAutoScalingCard` (lazy orchestrator).
- **Card ownership normalized**: the Replicas card's `BAICard` / error boundary moved from the page into `DeploymentReplicasCard`, so all sections own their own card and the page just renders sibling cards.
- **List fetch-policy normalized**: `DeploymentReplicasCard` and `DeploymentAccessTokensCard` now mirror `DeploymentRevisionHistoryTab` — `fetchPolicy: isInitialFetch ? 'store-and-network' : 'network-only'` (cached-first on the initial load, fresh on refresh / add-revision), all keyed via `useFetchKey()` + `INITIAL_FETCH_KEY`.
- Persistence keys (`table_column_overrides.*`) kept at their original values so users don't lose saved column settings.

## Verification

- `bash scripts/verify.sh` → Relay / Lint / Format / TypeScript pass.
- Runtime: detail page renders all section cards; each tab's refresh / filter / sort / pagination works (via `deployment(id:)`).

[BA-6569]: https://lablup.atlassian.net/browse/BA-6569?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
